refactor: use balance from new architecture
